# docs/api.yml
openapi: 3.0.3
info:
  title: Todolist API
  description: The APi for our todolist
  version: 1.0.0
servers:
  - url: http://localhost:3000
tags:
  - name: tasks
    description: Everything about Tasks
  - name: users
    description: Everything about Users
paths:
  /tasks:
    get:
      tags:
        - tasks
      summary: All tasks
      description: Return all the tasks for all the users
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      tags:
        - tasks
      summary: Create a new task
      requestBody:
        description: Should contain the new task informations
        content:
          application/json:
            schema:
              type : object
              example: {"label" : 'My task'}
      description: Create a new task and return it
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Bad Request
        "500":
          description: Internal Error
  /tasks/{id}:
    get:
      tags:
        - tasks
      summary: One task
      description: FInd a task by id
      parameters:
        - name: id
          in: path
          description: ID of tasks to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        label:
          type: string
          example: Coder une API REST avec nodeJs et Express
        done:
          type: boolean
          example: true
        userId:
          type: integer
          format: int64
          example: 1